example/untrainable_v1_2/requirements.txt
numpy>=0.16.0
spicy
matplotlib
pandas>=1.0.5
pycryptodome
boto3>=1.10.17
scikit-learn>=0.21.0
seaborn
cryptography
flatBuffers
attrs
cattrs
msgpack-python
jsonlines
aenum
Statsmodels
ruptures
git+https://github.com/plasticanne/TimeSeriesDataAnalysisLib.git